+++ Classes/Service/HugoExportService.php
@@ -26,14 +26,12 @@ namespace SourceBroker\Hugo\Service;
 
 use SourceBroker\Hugo\Configuration\Configurator;
 use SourceBroker\Hugo\Traversing\TreeTraverser;
-use TYPO3\CMS\Core\Database\ConnectionPool;
-use TYPO3\CMS\Core\Database\Query\Restriction\DeletedRestriction;
-use TYPO3\CMS\Core\Database\Query\Restriction\HiddenRestriction;
 use TYPO3\CMS\Core\Locking\Exception\LockAcquireWouldBlockException;
 use TYPO3\CMS\Core\Locking\LockFactory;
 use TYPO3\CMS\Core\Locking\LockingStrategyInterface;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Object\ObjectManager;
+use \SourceBroker\Hugo\Domain\Repository\Typo3PageRepository;
 
 /**
  * Class HugoExportService
@@ -49,7 +47,8 @@ class HugoExportService
      */
     public function exportTree(): bool
     {
-        $lockFactory = GeneralUtility::makeInstance(LockFactory::class);
+        $objectManager = GeneralUtility::makeInstance(ObjectManager::class);
+        $lockFactory = $objectManager->get(LockFactory::class);
         $locker = $lockFactory->createLocker(
             'hugo',
             LockingStrategyInterface::LOCK_CAPABILITY_SHARED | LockingStrategyInterface::LOCK_CAPABILITY_NOBLOCK
@@ -66,28 +65,25 @@ class HugoExportService
             }
         } while (true);
 
-        $objectManager = GeneralUtility::makeInstance(ObjectManager::class);
-
-        foreach ($this->getSiteRoots() as $siteRoot) {
-            $config = $objectManager->get(Configurator::class, null, $siteRoot['uid']);
-            if ($config->getOption('enable')) {
+        foreach (($objectManager->get(Typo3PageRepository::class))->getSiteRootPages() as $siteRoot) {
+            $hugoConfigForRootSite = $objectManager->get(Configurator::class, null, $siteRoot['uid']);
+            if ($hugoConfigForRootSite->getOption('enable')) {
                 // EXPORT CONTENT FROM TYPO3
-                /** @var $config Configurator */
-                $writer = GeneralUtility::makeInstance($config->getOption('writer.class'));
+                /** @var $hugoConfigForRootSite Configurator */
+                $writer = $objectManager->get($hugoConfigForRootSite->getOption('writer.class'));
                 $treeTraverser = $objectManager->get(TreeTraverser::class);
-                $writer->setRootPath($config->getOption('writer.path.content'));
-                $writer->setExcludeCleaningFolders([$config->getOption('writer.path.media')]);
+                $writer->setRootPath($hugoConfigForRootSite->getOption('writer.path.content'));
+                $writer->setExcludeCleaningFolders([$hugoConfigForRootSite->getOption('writer.path.media')]);
                 $treeTraverser->setWriter($writer);
                 $treeTraverser->start($siteRoot['uid'], []);
 
                 // USE CONTENT EXPORTED FROM TYPO3 TO RUN HUGO BUILD
-                $hugoPathBinary = $config->getOption('hugo.path.binary');
+                $hugoPathBinary = $hugoConfigForRootSite->getOption('hugo.path.binary');
                 if (!empty($hugoPathBinary)) {
-                    exec($hugoPathBinary . ' ' . str_replace(['{PATH_site}'],[PATH_site],$config->getOption('hugo.command')));
+                    exec($hugoPathBinary . ' ' . str_replace(['{PATH_site}'],[PATH_site],$hugoConfigForRootSite->getOption('hugo.command')));
                 } else {
                     throw new \Exception('Can not find hugo binary');
                 }
-
                 if ($locked) {
                     $locker->release();
                     $locker->destroy();
@@ -96,30 +92,5 @@ class HugoExportService
             }
         }
 
-    }
-
-    /**
-     * @return array
-     */
-    public function getSiteRoots()
-    {
-        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)
-            ->getQueryBuilderForTable('pages');
-
-        $queryBuilder
-            ->getRestrictions()
-            ->removeAll()
-            ->add(GeneralUtility::makeInstance(DeletedRestriction::class))
-            ->add(GeneralUtility::makeInstance(HiddenRestriction::class));
-
-        return $queryBuilder
-            ->select('uid')
-            ->from('pages')
-            ->where(
-                $queryBuilder->expr()->eq('is_siteroot', $queryBuilder->createNamedParameter(1, \PDO::PARAM_INT))
-            )
-            ->orderBy('sorting')
-            ->execute()
-            ->fetchAll();
     }
 }