+++ Classes/Command/HugoCommandController.php
@@ -15,7 +15,9 @@ namespace SourceBroker\Hugo\Command;
  * The TYPO3 project - inspiring people to share!
  */
 
-use SourceBroker\Hugo\Service\HugoExportService;
+use SourceBroker\Hugo\Service\HugoExportContentService;
+use SourceBroker\Hugo\Service\HugoExportMediaService;
+use SourceBroker\Hugo\Service\HugoExportPageService;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Mvc\Controller\CommandController;
 
@@ -24,15 +26,62 @@ use TYPO3\CMS\Extbase\Mvc\Controller\CommandController;
  */
 class HugoCommandController extends CommandController
 {
+
     /**
-     * Export content to hugo
+     * Export all
      */
     public function exportCommand()
     {
-        $hugoPageService = GeneralUtility::makeInstance(HugoExportService::class);
-        $this->outputLine('Generating hugo pages for all TYPO3 tree roots.');
+        $this->outputLine('Generating pages / content / media for all TYPO3 tree roots.');
+        if (
+            (GeneralUtility::makeInstance(HugoExportContentService::class))->exportAll()
+            && (GeneralUtility::makeInstance(HugoExportMediaService::class))->exportAll()
+            && (GeneralUtility::makeInstance(HugoExportPageService::class))->exportAll()
+        ) {
+            $this->outputLine('Success.');
+        } else {
+            $this->outputLine('Fail.');
+        }
+    }
 
-        if ($hugoPageService->exportTree()) {
+    /**
+     * Export pages
+     */
+    public function exportPagesCommand()
+    {
+        $this->outputLine('Generating Hugo pages for all TYPO3 tree roots.');
+        if ((GeneralUtility::makeInstance(HugoExportPageService::class))->exportAll()) {
+            $this->outputLine('Success.');
+        } else {
+            $this->outputLine('Fail.');
+        }
+    }
+
+    /**
+     * Export content
+     */
+    public function exportContentCommand()
+    {
+        $hugoExportContentService = GeneralUtility::makeInstance(HugoExportContentService::class);
+        $this->outputLine('Generating Hugo content for all TYPO3 tree roots.');
+
+        if ($hugoExportContentService->exportAll()) {
+            $this->outputLine('Success.');
+        } else {
+            $this->outputLine('Fail.');
+        }
+    }
+
+
+    /**
+     * Export media
+     */
+    public function exportMediaCommand()
+    {
+        $hugoExportMediaService = GeneralUtility::makeInstance(HugoExportMediaService::class);
+        $this->outputLine('Generating Hugo media for all TYPO3 tree roots.');
+
+        if ($hugoExportMediaService->exportAll()) {
             $this->outputLine('Success.');
         } else {
             $this->outputLine('Fail.');