+++ Classes/Domain/Repository/Typo3ContentRepository.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace SourceBroker\Hugo\Domain\Repository;
+
+use TYPO3\CMS\Core\Database\ConnectionPool;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+
+class Typo3ContentRepository
+{
+    /**
+     * @return array
+     */
+    public function getAll()
+    {
+        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)
+            ->getQueryBuilderForTable('tt_content');
+        return $queryBuilder
+            ->select('*')
+            ->from('tt_content')
+            ->orderBy('sorting')
+            ->execute()
+            ->fetchAll();
+    }
+
+}