+++ Classes/ContentElement/HeaderContentElement.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace SourceBroker\Hugo\ContentElement;
+
+class HeaderContentElement extends AbstractContentElement
+{
+    /**
+     * @param array $contentElementRawData
+     * @return array
+     */
+    public function getSpecificContentElementData(array $contentElementRawData): array
+    {
+        return [];
+    }
+}